FIX BUG IN users SCHEMA
Altered the users schema by adding password, email, password, first_name, last_name to the Columns and made all of them String(128) and not nullable. Observed that once a table has been created and the class is mutated it won't take effect in the database. The only Options where to drop the tables and re-run the code or alter the table schema